Rewrite Session to use SessionID instead of username
=Session struct.
add SessionType to specify the user is student or not.
add SessionID to take place of username as primary key.

=StartSession().
make sid inside and add the session to SessionMap.

+MakeSessionID
implement GUID to specifiy any session whether it is student or not

+SetCookies()
collect the set cookies steps to one function

+SessionTypeMask int
to enum Student =iota and then Company.

=lots of changes in companyLogin.go as there are several bug and wrong logic.